OptionalArgChecks

Provides two macros, @argcheck and @skipargcheck which give users control over whether to skip argument checking for better performance.

API

OptionalArgChecks.@argcheck

OptionalArgChecks.@skipargcheck

OptionalArgChecks.@argcheck — Macro

@argcheck ex

Marks ex as an optional argument check, so when a function is called via @skipargcheck, ex will be omitted.

julia> function half(x::Integer)
           @argcheck iseven(x) || throw(DomainError(x, "x has to be an even number"))
           return x ÷ 2
       end
half (generic function with 1 method)

julia> half(4)
2

julia> half(3)
ERROR: DomainError with 3:
x has to be an even number
[...]

julia> @skipargcheck half(3)
1

OptionalArgChecks.@skipargcheck — Macro

@skipargcheck ex

For every function call in ex, expressions wrapped in @argcheck get omitted recursively.